// src/enum/env.enum.ts
export enum Env {
  APP_NAME = 'app_name',
  HOST = 'host',
  PORT = 'port',
  LOGGER = 'logger',
  DB = 'db',
  DB_TYPE = 'db.type',
  DB_HOST = 'db.host',
  DB_PORT = 'db.port',
  DB_USERNAME = 'db.username',
  DB_PASSWORD = 'db.password',
  DB_DATABASE = 'db.database',
  DB_SYNCHRONIZE = 'db.synchronize',
  DB_MIGRATIONS_RUN = 'db.migrationsRun',
  PINO = 'pino',
  PINO_LEVEL = 'pino.level',
  PINO_IS_ROLL = 'pino.isRoll',
  REDIS = 'redis',
  REDIS_HOST = 'redis.host',
  REDIS_PORT = 'redis.port',
  REDIS_PASSWORD = 'redis.password',
  REDIS_DB = 'redis.db',
  JWT = 'jwt',
  JWT_SECRET = 'jwt.accessSecret',
  JWT_EXPIRES_IN = 'jwt.expiresIn',
  JWT_REFRESH_SECRET = 'jwt.refreshSecret',
  JWT_REFRESH_EXPIRES_IN = 'jwt.refreshExpiresIn',
  COOKIES = 'cookies',
  COOKIES_REFRESH_TOKEN_NAME = 'cookies.refreshTokenName',
}